Implement v3 compatibility layer as mixin on Controller class
